+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.0] - 2022-10-28
+  - Add support for indexed registers
+  - Add schemas for triggers
+  - Fix bugs in WARL class
+  - Add function to get a legal value from WARL fields.
+
 ## [3.4.0] - 2022-10-13
   - support x-extension parsing in ISA string
   - added check to capture duplicates in the extensions present in the isa string